Add skeleton for JavaScript rule parser
