+++ graylog2-server/src/main/java/org/graylog/plugins/pipelineprocessor/parser/RuleContentType.java
@@ -1,0 +1,32 @@
+/*
+ * Copyright (C) 2020 Graylog, Inc.
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the Server Side Public License, version 1,
+ * as published by MongoDB, Inc.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+ * Server Side Public License for more details.
+ *
+ * You should have received a copy of the Server Side Public License
+ * along with this program. If not, see
+ * <http://www.mongodb.com/licensing/server-side-public-license>.
+ */
+package org.graylog.plugins.pipelineprocessor.parser;
+
+public enum RuleContentType {
+    GL_PIPELINE_LANGUAGE("application/x-graylog-pipeline-language"),
+    JAVASCRIPT_MODULE("application/javascript+module");
+
+    private final String mimeType;
+
+    RuleContentType(String mimeType) {
+        this.mimeType = mimeType;
+    }
+
+    public String mimeType() {
+        return mimeType;
+    }
+}
